nearbyshare: clear transferred files at the end of Phone<->CB tests
Clear the contents of both devices' Downloads directories in PostTest to
clean up any test files. This will help prevent false positives when
running the tests locally with the keepState option (CL:2863914) and
when we eventually nest the fixtures to limit the # of logins
(CL:2786383). Without this, we can get in the situation where a device
receives file.ext, and then receives it again in a later test, where it
will be renamed file (1).ext or something like that. When we perform the
file hash comparison, though, we will use the filename from the test
data (file.ext) so we'll be checking the wrong file.

BUG=chromium:1184352
TEST=tast run <dut> nearbyshare.*Phone*

Change-Id: I15a99b92553b1fac68a1e0d628adda102149ab14