cuj: Use fixture without KeepState
Change CUJ test fixture where KeepState is not enabled. This is a
workaround to clear the DUT ownership and make sure the new login account
is the DUT owner. This is needed when executing the CUJ test with a login
account that has never done login before.
Also add "arc" dependency for tab_switch_cuj2 case because the fixture
enables ARC.

BUG=chromium:1207756
TEST=fast_build.sh -T
TEST=tast run <vars> <DUT> QuickCheckCUJ2.*
TEST=tast run <vars> <DUT> TabSwitchCUJ2.*

Change-Id: I1e9e3573e981bac8a09c0bc2782ac349d3734e41