cuj: CUJ fixture and dependency fixes
Change CUJ tests to use fixture with KeepState enabled. There was a bug
(see issue 1202797) and a workaround was introduced
to clear the DUT ownership by removing KeepState and make sure
the new login account is the DUT owner. This is no longer needed because
bug was not reproduceable on R92.
Also add "arc" dependency for tab_switch_cuj2 case because the fixture
enables ARC.

BUG=b:187795917
TEST=fast_build.sh -T
TEST=tast run <vars> <DUT> QuickCheckCUJ2.*
TEST=tast run <vars> <DUT> TabSwitchCUJ2.*

Change-Id: I1e9e3573e981bac8a09c0bc2782ac349d3734e41