arc: Fix ureadahed is generated almost empty
It occurs that tracing now is reset on Chrome session starts. That makes
it incompatible with our approach to start ureadahead tracing before
Chrome stars and log until ARC is opted in. Looks like this is
regression. The fix now to start tracinfg before explicit ARC optin
after Chrome is already signed-in. Also this fixes another issue,
setting the filter for optimization. This was done for VM however does
not work for container where entries appear in tracing log relative to
root.

BUG=b:184804045
TEST=tast run ${DUT} arc.DataCollector.[vm_]local
     Note, testing was done over the PFQ eve build and test image kohaku
     Confirmed that ureadahead pack contains reasonable volume 400+mb.

Change-Id: Iba30acb11f757a217166f4cb06e1740fb6f5357e