security.Mtab: Simplify logic and baseline (reland).
This test had become too brittle, and it was failing for reasons that
didn't imply a security bug.

Refactor the test to only complain if new mounts show up *that would
present a security bug*. These would be mounts that don't include all
of the "nodev", "noexec", "nosuid", flags; or include those but also
widen access to the mount by using "gid=" or "mode=" flags.

This change will make it so that fewer folks have to update the
baseline when adding an innocuous mount, and it also cuts down on the
existing size of the baseline, reducing it to less than half of what it
was.

The first time this CL landed devices with 3.18 kernels failed in
postsubmit:

Unexpected mount /run/arc/debugfs/tracing in /etc/mtab with
device "debugfs", type "debugfs",
options [rw seclabel nosuid nodev noexec relatime gid=605 mode=750]

Fix that by re-adding the baseline expectation for
/run/arc/debugfs/tracing. The problem is that this mount shows up
with different mount type (debugfs or tracing) and gid/mode depending
on kernel versions.

BUG=chromium:1206687
TEST=Passes on betty-pi-arc.

Change-Id: I6ee80d4bfaad576419fcf4f39a1369618ef89d5f