security.Mtab: Simplify logic and baseline (reland).
This test had become too brittle, and it was failing for reasons that
didn't imply a security bug.

Refactor the test to only complain if new mounts show up *that would
present a security bug*. These would be mounts that don't include all
of the "nodev", "noexec", "nosuid", flags; or include those but also
widen access to the mount by using "gid=" or "mode=" flags.

This change will make it so that fewer folks have to update the
baseline when adding an innocuous mount, and it also cuts down on the
existing size of the baseline, reducing it to less than half of what it
was.

BUG=chromium:1206687
TEST=Passes on betty-pi-arc.

Change-Id: I6ee80d4bfaad576419fcf4f39a1369618ef89d5f