tast: Use linuxssh instead of rsync
Rsync is failing to copy files to the DUT, use linuxssh.PutFiles
instead. Also not all of the files exist on every dut, so the chmod was
failing. Changed to verify that the files exist before trying to act on
them.

BUG=None
TEST=sudo emerge chromeos-base/tast-cmd && \
tast run --var=servo=localhost:9999 192.168.1.61 firmware.BootMode*

Cq-Depend: chromium:2885285
Change-Id: I861545382f906a9a839081f61fbe611b07689507