wifi.UdevRename: exclude asurada from critical variant
Asurada has known issues that will not be fixed promptly (b/183992356).
Let's exclude it from mainline, so it doesn't pollute other tests
running on Asurada (that won't find a WiFi device).

While we're at it, move Elm/Hana to a software dependency, so we can
still run on kernel 4.19 where the issue is fixed.

BUG=b:183992356, b:187682828
TEST=wifi.UdevRename, wifi.UdevRename.informational on asurada, others

Change-Id: I94d3b731bb286819476f6fbf7f67c558a3019fce