crostini: Add tast test for restoring backup to a resized container
This automates the steps described in the bug. The test performs
operations via Settings and uses the terminal app to run a command
that verfies the md5sum of /home, /usr, and /etc are the same on
restore to a resized disk as before the backup was taken.

BUG=chromium:1205743
TEST=tast run IP crostini.ResizeBackupRestore.*

Change-Id: I55e5bf3a8830b51d4097f98d23dbfecb894bfa5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2888848
Tested-by: Nicholas Verne <nverne@chromium.org>
Auto-Submit: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Judy Wu <jinrongwu@google.com>
Reviewed-by: David Munro <davidmunro@google.com>
Commit-Queue: Nicholas Verne <nverne@chromium.org>