tast: Handle unexpected errors, omit setpowerstate
Although we expect the ssh to the DUT to run poweroff to fail, it is
only expected to timeout. Other errors should still be reported as fail.

Servo.SetPowerState either sends an EC command to power off or holds down the
power button to make sure the DUT is off, since we just did a
WaitUnreachable, we know it is off and don't need to turn it off again.

BUG=None
TEST=sudo emerge chromeos-base/tast-cmd && \
tast run --var=servo=localhost:9999 $DUT firmware.BootMode*

Cq-Depend: chromium:2888885
Change-Id: I44751ae092d08d41ee355ed052470075df41053e