tast: RoamFT: use BSS TM Request instead of D-Bus RequestRoam
MBO test plan requires that we test FT with BSS TM Request, both MBO
features. Also, it makes more sense to use BSS TM Request anyway since
it's actually a standardized way to induce a roam, as opposed to the
D-Bus roam request.

Note that as it stands, we won't ever actually use BSS TM Request to
roam since DiscoverBSSID induces a roam based on signal level (see
b:186800736). $TEST was run with both APs on 11n to avoid hitting this
bug and actually test the roam request as well.

BUG=b:180664307,b:174963631
TEST=run RoamFT.* on hatch

Change-Id: Ib085e4098f3cfab952b617dda2cc8783240d58ad