tast-tests: fix ui.WindowArrangementCUJ.clamshell_mode failures in lab
It happens often in lab that there are more than 2 desk mini views
which might be caused by former tests ran on the device before.
As long as there are >=2 mini views, the operation of dragging
the second window to another desk will succeed.

BUG=chromium:1208312
TEST=tast run $DUT ui.WindowArrangementCUJ.clamshell_mode

Change-Id: I466979d99e23ac2ad0cd46693a45fe954e663a53
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2891809
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Patrick Meiring <meiring@google.com>
Tested-by: Yichen Zhou <yichenz@chromium.org>
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>