network.DiagFailFirewall: save and restore iptables
This change saves the initial state of the iptables and restores them
when the test is complete. Resetting shill is not enough to restore the
iptables to their original state. This was causing issues for other
tests since network traffic was blocked.

BUG=chromium:1203018
TEST=tast run hatch-dut network.DiagFailFirewall* (pass)
(dut) iptables -S (iptables do not show blocked 80 or 443 ports)

Change-Id: Ie4980c2ee090ecf6d7f10145f324692559fb58b6