tast-tests: Add support for windowState property on ARC-R
This CL adds the support for providing TaskInfo.windowState property on ARC-R, which blocks some tests to be enabled on ARC R as *_vm.
In P, “dumpsys activity activities” contains the windowState information.
In R, however, it does not, and the windowState information is in the “dumpsys Wayland” instead.
Therefore, in this CL, we
* support windowState property on ARC-R by merging the results of “dumpsys activity activities” and “dumpsys Wayland”
* rename the DumpsysActivityActivities to TaskInfosFromDumpsys as the function no longer directly returns the output of “dumpsys activity activities” on ARC-R.

BUG=b:152576355
BUG=b:168011276
BUG=b:179525662
TEST=run arc.SplitView.* with android_vm option
TEST=run arc.WindowState.*_vm

Change-Id: I3060bd5ab50f5609bd5e0ea00cf1815e9ff3e5b4