Lacros: Fix Lacros tests expectations
Context: Lacros used to show a 'Welcome to Chrome' upon first run, but
it has recently changed the behavior to always show a New Tab Page.

Lacros Tast tests are broken at ToT due to the changes, but it's not
noticed at OS side because it's running a month-old Lacros, so this
CL fixes the test expectations.

BUG=chromium:1208845
TEST=tast run <dut> lacros.ShelfLaunch lacros.AppLauncherLaunch

Change-Id: Ifcc9dd111cc41cc567e8aa467d22e256256a37d1