crash: Only run certain tests on internal builds.
In the linked bug, crash sender tests (and some other tests) broke
because we started including "chromiumos" in the os-release file, which
crash_sender doesn't expect.

We don't actually care about crash_sender's behavior on these images,
and in future we'll stop sending crashes on them, so disable tests there.

BUG=chromium:1208870
TEST=tast run $DUT crash.Sender* crash.Serializer* crash.User*

Change-Id: I584cbbd239859b551c187c75100ab5866ca43d1d