tast-tests: disable long running tests on vm in NNAPIDriver
This CL disables long running tests for VM platforms and removes the
informational flag.

NNAPIDriver.{cts,vts_1_2,vts_1_3} take a long time to run in VM's. We
get a basic smoke test from the vts_1_0 and vts_1_1 which run in a few
seconds, so we'll leave them enabled for VM's.

BUG=b:182254380
TEST=tast -verbose run <amd64-generic VM>  platform.NNAPIDriver.*

> platform.NNAPIDriver.cts      [ SKIP ]
> platform.NNAPIDriver.vts_1_2  [ SKIP ]
> platform.NNAPIDriver.vts_1_3  [ SKIP ]
> platform.NNAPIDriver.vts_1_0  [ PASS ]
> platform.NNAPIDriver.vts_1_1  [ PASS ]
> Ran 5 local test(s) in 2.484s

Change-Id: I37a7afc17dac20a3e816ef6aac2290a55ce0b043