Lacros: Update Lacros test artifacts path
/mnt/stateful_partition should not be used to execute binaries, and it
is recommended to use /usr/local, so this CL updates the test artifacts
to /usr/local.

This CL also refactors a few variable names to improve readability.

Additionally, it is suspected that the flakiness of lacros.Basic
might be due to executing binaries in /mnt/stateful_partition and this
CL might fix the problem.

BUG=chromium:1198252
TEST=tast run localhost:2222 lacros.Basic

Change-Id: Ib8286685cb5c101a8981b8cd25aa6c8fa9859d45