tast-tests: Create test PWA for Play Billing integration
Play Billing integrates with PWAs surfaced on the Play Store as of M89.
This test PWA will be served locally and a private listing on the Play
Store will be used to point to this app. Follow up CLs will provide
wrapper functions that interact with the JS on the page.

BUG=chromium:1209704
TEST=./fast_build.sh -T

Change-Id: I5d6322915644cc99f34ab3c1c48d1cd749597d5f