tast-tests: Create test PWA for Play Billing integration
Play Billing integrates with PWAs surfaced on the Play Store as of M89.
This test PWA will be served locally and a private listing on the Play
Store will be used to point to this app. Follow up CLs will provide
wrapper functions that interact with the JS on the page.

The APK has been published onto the Play Store and exposed to a rhea
test account. Using this GCP bucket flow for now until we have the
source code loaded into the android/ directory. This APK has been
generated by bubblewrap-cli and points to 127.0.0.1:80 from where the
PWA will be served.

BUG=chromium:1209704
TEST=./fast_build.sh -T

Change-Id: I5d6322915644cc99f34ab3c1c48d1cd749597d5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2896822
Tested-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Jeevan Shikaram <jshikaram@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>