shared_filesystem_state: make default list and add logging
Make the a base list of expected mounts and use as default if ARC isn't
available on the device. Add /run/netns to the list of defaults as it
seems to show up on most systems and is apparently used for non-ARC
purposes. Also add the first unexpected shared mount to the error message for
better visibility in tastboard.

BUG=None
TEST=run tast test on atlas

Change-Id: Ia41ebb1f24c3adb2f70715ce3c349c2097a412e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2903128
Tested-by: Nicole Anderson-Au <nvaa@google.com>
Commit-Queue: Nicole Anderson-Au <nvaa@google.com>
Reviewed-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>