network: VPNConnect: Use ConnectNamespace
In the tast test for vpn, the vpn server is established inside a netns,
and currently we create that netns and a pair of veth,  and allocate ips
in the test code.

This patch change this logic to doing it by calling the ConnectNamespace
api provided by patchpanel, and thus we don't need to maintain the netns
in the test code.

BUG=b:185210339
TEST=tast run $DUT network.VPNConnect.l2tp_ipsec_psk

Change-Id: Ifa69dbf1f73ec88859e88806beb5e9f5ded9b6bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2903411
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Mohamed Omar <mohamedaomar@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>