scanapp: Wait for idle virtual printer at startup
When the virtual USB printer is first "plugged in", it can take several
seconds for all the enumeration and probing to finish.  If the app is
started before this finishes, it can see errors like a missing scanner
or scanning timeouts.  To eliminate this race:

1.  Wait for the ippusb_bridge socket to appear.  This approximates the
    end of USB enumeration.
2.  Wait for CUPS to finish probing the printer.
3.  Launch the scan app after the previous two steps finish.

Also remove the calls to kill ippusb_bridge, since these haven't been
needed for a while.

BUG=b:188591629, b:188590605
TEST=150 iterations of tast run scanapp.*

Change-Id: I4c2a58326ddde7ebeae21c0f41c991ff30927d5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2912296
Reviewed-by: Paul Moy <pmoy@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>