tast-tests: match probe components only by cid
Some fields are mutable (e.g. firmware) which leads the probed
components to a different one, and causes the tests to be flaky.  This
CL tries to remove the "_{qid}" suffix in the component names where
`{category}_{cid}` should be sufficient to verify the probe works as
expected.

Note that this only applies for component names following the
AVL-embedded name policy `{category}_{cid}_{qid}` with optional `#{seq}`
suffix.  For other cases, they are unaffected by this CL.

BUG=b:188627930
TEST=fast_build.sh -T

Change-Id: If96ccc68b41efff43e2625531380123450291435