crostini.NestedVM: Surface test program errors better
Currently the output of the test program is just dumped in the logs if
it fails. This makes it hard to tell if two different runs failed for
the same reason. Fix this by adding the program output to the error
message surfaced in stainless.

BUG=chromium:1205703
TEST=Made test fail

Change-Id: I2fe3d5df08fb23f28f94a1a582781d0a98adcee7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2908885
Tested-by: Fergus Dall <sidereal@google.com>
Auto-Submit: Fergus Dall <sidereal@google.com>
Reviewed-by: Judy Wu <jinrongwu@google.com>
Commit-Queue: Judy Wu <jinrongwu@google.com>