kernel: Add a test to confirm console-ramoops works across reboot
There have been some cases where console-ramoops doesn't seem to exist
after a reboot. Let's make sure that a warm reboot, via 'reboot' on the
DUT, creates a ramoops file in /sys/fs/pstore/console-ramoops-0 that has
the contents of the previous kernel boot. We write a message into the
kernel logs to search for on the next boot to make sure this worked
properly.

BUG=b:188926982
TEST=tast run $DUT kernel.PstoreConsoleRamoops

Change-Id: I510edc47adba89bd88aa78941b6278c2cda16b3c