Add feature to find a UI node using regex
There is a need to locate a node using regex matches for attributes like classname. Currently, it is only able to do so for name attribute.

Add support to generate JavaScript code for attributes with values of regex type.

Add ClassNameRegex method to find a node with a class name matching a regular expression.

BUG=b:188881833
TEST=./fast_build.sh -t chromiumos/tast/local/chrome/uiauto/nodewith

Change-Id: I3a0c0f49e15f6a51e261d40946605db4c19d4459