video.Contents: Offset outer corner samples.
This CL offsets three of the outer corner sampling points (top-left,
top-right, and bottom-left) in order to tolerate color blending
artifacts that are expected along some of the edges in the "exotic crop"
cases. The bottom-right corner is left as-is because we never expect
blending artifacts there.

The goal is to bring a lot of data points to a color distance of less
than 100 (see https://screenshot.googleplex.com/Bc5tv3UtK7pXAgs for a
current distribution).

BUG=b:159158467
TEST=video.Contents.h264_360p_exotic_crop_composited_hw on eldrid

Change-Id: Ibf28ee949f520e6e6ae3cc82f85c6f71b876254a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2913305
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>