video.DrawOnCanvas: Pass a reference image.
This CL allows video.DrawOnCanvas to use a reference PNG file which
contains the expected result of drawing the video onto a canvas. For
now, we only use the reference file to set the size of the canvas equal
to the size of the reference image. The reason is that we'd rather draw
the video on the canvas without any scaling in order to prevent
filtering artifacts from interfering with our color checks.

We'll use the reference image for color checks in a follow-up CL.

BUG=b:159158467
TEST=video.DrawOnCanvas.h264_360p_hw on eldrid

Change-Id: Ia47966f4c260077a86d7cd7c3fbd3f43f0d0a910