tast-tests: skip known concurrent D-Bus call at boot
The D-Bus methods of runtime-probe are designed to be one-off, so only
the first message of concurrent D-Bus calls will be processed and the
daemon will stop.  In CL:2893765 we extend the sleep time of init script
of hardware-verifier with 20 seconds, which results in concurrent D-Bus
calls (ProbeCategories method triggered by hardware-verifier at boot and
GetKnownComponents triggered by tast tests) and makes tests flaky.  This
CL by default makes all D-Bus calls retry once if the first attempt
fails to avoid this issue.

BUG=b:189072439
TEST=manual test

Change-Id: Idf41edda66e626af437b49c10a86ad1a41603e55