adb: fix RemoveContents deleting empty dirs
The RemoveContents function deletes empty directories, since ListContents
returns a slice with a single empty string if the folder is empty due to
how strings.Split works. This is causing some flakiness in Nearby tests
as it causes /sdcard/Download to be deleted sometimes. Now ListContents
will just return nil if the target directory is empty.

BUG=b:187795466
TEST=tast run <dut> nearbyshare.*

Change-Id: I5dc781b06d8fb296768c5013d4272fc661e13fa8