Stabilize handwriting tests on longfom canvas
When for the first time using the handwriting on a longform canvas, an information dialogue shows up. Current handwriting tests do not handle this and results in test failures for IME using longform. Plus, clearing a longform canvas is also different from clearning a legacy canvas.

Add support to close the information dialogue if one shows up.

Add a cleaning canvas function to handle both a longform canvas and a legacy canvas.

BUG=b:187766993, b:188489307, chromium:1175982
TEST=tast run <test-device-ip> inputs.VirtualKeyboardHandwring*

Change-Id: I74db5eb858925ee9f7f95fbeea77efaab88c151e