tast-tests: Unflake spoken feedback tests.
This change unflakes both the Google TTS and eSpeak
spoken feedback tests by:

1. Creating a more robust algorithm for connecting
to a TTS background page. The old method was to lazily
pick the first background page out of multiple targets, but that
has proven to be unreliable. The new algorithm iterates through
each target, sends a test utterance from ChromeVox, then waits for
it to appear in the target background page. If data flows from
ChromeVox to the target background page, then we know it's the one
we should use for tests.
2. eSpeak was recently updated [1], which caused the eSpeak test
to fail due to registering a listener on the wrong function.
We now register a listener for onSpeakWithAudioStream, which is
the function that gets called whenever an utterance is spoken by
eSpeak.
3. Performs cleanup and consolidation where possible.

[1] https://crrev.com/c/2792065

BUG=chromium:1179334, chromium:1179338
TEST=a11y.SpokenFeedback.*

Change-Id: Ib01b1e326a45d1580aec31471e157030c96d5b5b