network: VPNConnect: Add basic wireguard test
This patch adds two tests for wireguard:
- network.VPNConnect.wireguard
- network.VPNConnect.wireguard_psk
The test will create and configure a wg interface in another netns as
server, invoke shill's D-Bus API to create a wg service, connect the
service to the server, and verify if the overlay address of the server
is ping-able.

Also changes GetLogContents() method in network_chroot so that it can
accept a list of paths (since different vpn servers have different
numbers of logs), and return an error if any file in the list does not
exist.

BUG=b:191830846
TEST=tast run $DUT network.VPNConnect.wireguard*

Cq-Depend: chromium:2981951
Change-Id: Ibc16d4075863c8c719d6085b2f48d2f3c4af9929