Only compute custom trace metrics in lacros.GpuCUJ for eve.
We found that it was taking a really long time in the lab to run. We
only need these stats temporarily, and only for eve. This CL uses
lsb-release to only compute these metrics on eve. We will remove this
custom metric and this code after a while.

BUG=chromium:1140906
TEST=tast run <dut> lacros.GpuCUJ.maximized

Cq-Depend: chromium:2919307, chromium:2919149
Change-Id: I14e6c05f85b1b61351d36a8089acf9338e7881f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2919149
Tested-by: Eliot Courtney <edcourtney@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Eliot Courtney <edcourtney@chromium.org>