tast-tests: mass update Vars to VarDeps
Update Vars to VarDeps for runtime variables that are required to run
the test.
I collected a list of strings that are used in s.RequiredVar() and never
in s.Var(), and moved the runtime variables in Vars matching the list
to VarDeps. (Creating the list, arguments that are not string literals
were manually examined.)

BUG=b:187795463
TEST=fast_build.sh -T

Change-Id: Id48412353ce24add91ee7c3aa57f6cb509f92857
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2919549
Tested-by: Keigo Oka <oka@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>