kernel: Add a test to measure bloat
Add a kernel test that measures bloat. Currently we measure the number
of drivers that aren't attached to any device, but builtin to the kernel
image, and devices that don't have any driver bound to them. The two
stats are recorded as a metric so we can track this over time.

BUG=b:189388451
TEST=tast run $DUT kernel.Bloat

Change-Id: I12bcaad9fa9654a211d7078258282deacb8084cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2920638
Tested-by: Stephen Boyd <swboyd@chromium.org>
Auto-Submit: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>