arc.VideoEncodeAccelPerf: Increase perf test timeout to 10 minutes.
The test timeout for the ARC++/ARCVM video encoder performance tests is
currently set to 4 minutes. But running the tests on slower end devices
can take up a lot more time than that, causing lots of flakiness in our
performance tests across a wide range of devices. A test on zork shows
that the test can easily take up to 6 minutes on that device.

The test timeout used to be 5 minutes but was reduced to 4 minutes in
crrev.com/c/2713556, further contributing to the flakiness. This CL
sets the timeout to 10 minutes, so we're sure the test won't time out
on lower end devices. The stress on the lab caused by devices hanging
and using this maximum timeout should be pretty limited, as this only
seems to happen on very rare occassions. The vast majority of test
issues are crash-related.

BUG=b:169915030
TEST=arc.VideoEncodeAccelPerf.h264_1080p_i420_vm

Change-Id: I29b78c69fcdd1258e807c134ce7fe711bf143a9e