network: DNSProxy: Add DNS proxy test
DNS proxy tests check if DNS functionality are working
properly when DNS proxies are running. The tests check
4 scenarios:
1. DNS queries should be successful by default.
2. DNS queries should fail when the corresponding DNS
   query type (plaintext DNS or secure DNS) is blocked.
3. DNS queries should be successful on a VPN.
4. DNS queries should fail on VPN if DNS on the VPN
   server is blocked (confirms that the queries got
   routed through the VPN server).

The tests are done on system, user, chrome, Crostini, and
ARC. Because ARC by default uses DNS-over-TLS by default,
the negative tests will not be run for ARC.

The tests are also done with different DoH modes (Off,
Automatic, and Always On).

BUG=b:180551432
TEST=tast run ${DUT} network.DNSProxy.*
TEST=tast run ${DUT} network.DNSProxyOverVPN.*

Cq-Depend: chromium:2946901
Change-Id: I902efcf8c9280e26d41d50fc815d441532500d61