fix apps.LaunchHelpApp
Check the log on https://tastboard.corp.google.com/test/apps.LaunchHelpApp.clamshell_oobe_stable

The error message is the same "failed to find node with properties: {name: /^Explore$/, role: rootWebArea}"

In the ui_tree, the window node can be found but rootwebarea is not yet shown. So it supposed to be a rendering performance issue. Sincee Help app is launched in background during OOBE. So this latency can be ignored.

This issue can be reproduced via leased skylab device snappy, appears 3 / 100. Extending the timeout to 1 min in waiting app. It passes consistently.

BUG=b:189696971
TEST=tast run <DUT> apps.LaunchHelp*

Change-Id: I630651cfd5df862fe6247cf4be50e48191455a10