Fix flakiness caused by IsNodeFound
At present, IsNodeFound immediately checks if a node exists. It can cause flakiness in some boards (i.e. Kukui).

Add a method IsNodeFoundWithTimeout to check the existence of a node with polling. It repeatedly checks if a node exists before the  specified timeout.

Refactoring handwriting to use this function for detecting lonform and undo key.

Minor fix in vkb to make sure the infoDialogue is properly closed.

BUG=b:188489307, chromium:1175982
TEST=tast run <test-device-ip> inputs.VirtualKeyboardHandwring*

Change-Id: I24c581383eebe4ff6074b36b712834b66e24f353