Fix flakiness caused by IsNodeFound
At present, ClearCanvas uses IsNodeFound to check if the "undo" key
exists. It can cause flakiness in some boards (i.e. kukui) due to
the immediate check.

Use IfSuccessThen with WaitUntilExists to check and clean the canvas
instead of IsNodeFound. This fixes the issue caused by IsNodeFound.

Minor fix in vkb to make sure the infoDialogue is properly closed.

BUG=b:188489307, chromium:1175982
TEST=tast run <test-device-ip> inputs.VirtualKeyboardHandwring*

Change-Id: I24c581383eebe4ff6074b36b712834b66e24f353