firmware: Always print flash_fp_mcu command output
The target of this test is specifically to check flash_fp_mcu
on each platform. We should be printing all output to be able
to spot warnings. This is not necessary for other tests that
use flash_fp_mcu.

BUG=none
TEST=# On zork
     tast run -var=servo=localhost:9999 localhost:2222 \
       firmware.FpFlashFpMcuHello
     # Check flash_fp_mcu output is shown.

Change-Id: Icb2c126442875262dceb55f037a4d63110d8b92f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2927229
Tested-by: Craig Hesling <hesling@chromium.org>
Auto-Submit: Craig Hesling <hesling@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>
Commit-Queue: Craig Hesling <hesling@chromium.org>