[play-billing] Play billing custom fixture for setup
Upcoming Play Billing tests require setup of ARC coupled to a PWA that
is served. Add a fixture that is a child of arcBootedForPlayBilling
which performs the required setup. Left TODO's in there for future
implementation to provide a wrapper around the Play Billing PWA and
ensure SetUp and TearDown includes more robust functionality.

BUG=chromium:1210330
TEST=./fast_build.sh -T
Disallow-Recycled-Builds: test-failures

Change-Id: I937f74d811e538c24855a82d56cce56cea3a599b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2929601
Reviewed-by: Judy Wu <jinrongwu@google.com>
Reviewed-by: Jeevan Shikaram <jshikaram@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Tested-by: Ben Reich <benreich@chromium.org>