[play-billing] Add TestApp to interact with Play Store PWA
All the tests that rely on the playBillingFixture will use a test
Android application to exercise the play billing functionality. Create a
wrapper around the TestApp so that all the difficult functionality can
be abstracted into easier helper functions. Complete all that setup in
the fixture to ensure small concise tests are written.

BUG=chromium:1210330
TEST=./fast_build.sh -T

Change-Id: I2a8f41fbbf505f14638323a8437608a71f69d227