Skip on scarlet for ui.TabletOperations.
It appears that the test is failing on scarlet on CQs, and the
test fails due to incorrect display rotation handling on this
device.

This CL skips the test on scarlet to unblock the CQ.  I will
look into the problems later and re-allow it when it's fixed.

BUG=b:189704582
TEST=manually

Change-Id: Id7a9af6049eb511e459cb1f9cbc313a37a632c78
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2930638
Tested-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
Commit-Queue: Raul E Rangel <rrangel@chromium.org>
Owners-Override: Raul E Rangel <rrangel@chromium.org>