tast-tests: Remove most tablet_mode software deps.
Since tablet mode can be toggled as needed in tests and a virtual
touchscreen device is used for devices without a real touchscreen, the
tablet_mode software dep is most-commonly used to indicate the need for
an internal display (the virtual touchscreen requires a real display).

Migrate tests other than arcappcompat to use hwdep.InternalDisplay()
unless they already set that hwdep or have a more-specific one (e.g.
hwdep.TouchScreen() or filtering on a particular set of models).

BUG=b:189921066
TEST=CQ, tast run the affected tests on a dooly

Change-Id: I826c23a7e08e114851a14519586a3d42c14e8e33