tast-tests: Add lacros.MemoryStress
This test opens some synthetic pages to allocate a lot of JavaScript
objects to create memory pressure on Lacros . The test switches between
the pages and reload crashed pages. It checks the count of reloading
crashed pages.

The common parts of lacros.MemoryStress and platform.MemoryStressBasic
are moved to local/memory/stress/.

BUG=b:190130454
TEST=tast run $DUT_IP lacros.MemoryStress

Change-Id: I5341eae63f4b66945eb75e428c1ed01c5e278b9b