Lacros: Promote lacros.ShelfLaunch
This test looks stable enough as it has been passing reliably for 20+
consecutive runs, thus promote it to be critical to prevent
regressions. The only errors are on Zork boards, but it's a separate
issue unrelated to Lacros.

This only promotes this test for boards with the lacros_stable
variation. Currently, this is all non-VM boards that support lacros.

BUG=chromium:1174896
TEST=tast run localhost:2222 lacros.ShelfLaunch

Change-Id: I55b6663750b9c7c4c36f1781b30551901d2fe89e