tast-tests: Add retries while loading TOS in optin
Tast board shows that arc.Optin is stable and it has retries enabled
arc.Optin.unstable is the same test without retries and most of the
failures occur while loading terms of service page.

In this change we're going to add retries around loading TOS and if
the unstable and stable tests are showing same stability then retries
from optin will be removed and unstable will be promoted to stable.

BUG=b:189285287
TEST=tast run <DUT> arc.Optin.vm

Change-Id: I4a9c3929c0c97adf0713480d7cc9b374a11c0dc3