tast-tests: ChromeVox hint test
This change adds a test for the ChromeVox hint in OOBE. The test
confirms correct behavior by:

1. Showing the welcome screen.
2. Creating a SpeechMonitor, which can be used to monitor utterances
given by the TTS engine.
3. Ensuring that the dialog is shown and the announcement is
spoken after an idle period.

This change also removes the flag
--enable-oobe-chromevox-hint-timer-for-dev-mode from restart.go [1],
which was previously added to ensure this feature was enabled.
Now that we have an explicit test for this feature, we can remove
it for all other tast tests.

[1] https://source.chromium.org/chromiumos/chromiumos/codesearch/+/
main:src/platform/tast-tests/src/chromiumos/tast/local/chrome/
internal/setup/restart.go;l=63

BUG=b:187795625
TEST=a11y.ChromevoxHint

Change-Id: I83f2feefb61b84970039f504f77e4bb6ff4138b2