Stabilize VK voice input tests
Currently, the VK voice tests are experiencing flakiness due to
not receiving the audio playback properly.

Add a retry to play the audio multiple times to reduce flakiness.

Modify the SwitchToVoice functions to trigger voice layout via UI,
instead of calling background API, which was a temporary solution due
to limited time of the previous dev.

BUG=b:188489307
TEST=tast run <DUT> inputs.VirtualKeyboardSpeech*

Change-Id: Ic2283969cd2b8293ed0b0abc859a23e2f020b102