tast: Changed the servo spec to include ssh port
When connecting to a skylab machine, it is necessary to use ssh port
forwarding, and the ssh port for connecting to the servo host will not
be on port 22.

Changed the servo string to be of the form host:9999:ssh:22, and allow
for ipv6 addresses such as [::1]:9999:ssh:22, also the port and ssh port
can be omitted, and the entire servo string can be omitted, and it
defaults to localhost:9999:ssh:22.

BUG=None
TEST=Added unit tests, also connected to local DUT and skylab DUT.

Change-Id: Ib00f7af82c55eae1c316a6d010b9a5afa6d507d5