tast: Get/put a firmware image section remotely
NewRemoteImage is like NewImage, except that you must pass in a
ServoHostCommandRunner (i.e. servo proxy) and the programmer to pass to
flashrom, and a section, and it returns a firwmare image for that
section.

WriteRemoteFlashrom is like Image.WriteFlashrom, but uses the passed in
programmer and servo host. To save bandwidth and time, only the
specified section is copied to the servo host and flashed, so the
specific flashrom args used are different from WriteFlashrom().

BUG=b:185271203
TEST=Used in firmware.ServoGBBFlags

Change-Id: I88f4bd9dd05a978b41f8c78efcf3af6384e828a7