tast: Clear the ECUARTRegexp before and after cmds
Sometimes the ECUARTRegexp can be left with garbage in it from a prior
command, which might be from another test, or the dut-control command.
Changed to clear the regexp before every no-output command, and to reset
it after a command that returns output. This matches what Tauto does.

BUG=b:185271203
TEST=Ran firmware.ServoGBBFlags

Change-Id: I9de4a9648a30f2938693ed80f2d25b3e9ecf6579
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2945094
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Auto-Submit: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>