Resize PIP using activity.ResizeWindow and drag from outer corner
This changes fixes two sources of flakiness during PIP resize
1. Dragging from the inner corner relies on the PIP menu being
visible. Hovering and waiting is not always reliable.
2. In R, the inner corner drag handle icon sometimes disappears
even if the menu is visible, making it impossible to drag to
resize.

To fix both issues, this change resizes using the outer drag handle
of the window, just as normal windows are resized.

BUG=b:168011269
TEST=tast run my_crbook arc.PIPEnergyAndPower.big_vm (and big_blend_vm)

Change-Id: Ie9199358b7a94c9a63b84786defaafad1e8df975