quicksettings: add support for muting/unmuting microphone
Add functions for interacting with the microphone mute toggle button in
Quick Settings' audio settings. The functions let you check the current
state and mute/unmute the microphone as needed.

BUG=b:189749471
TEST=local modification to ui.QuickSettingsMicGainSlider to ensure the
new functions work as expected. They will be incorporated into a new
test in CL:2932160 that will test the button's integration with the
audio stack.

Change-Id: Idfa7817fc184326976d88d037144816d7b7a180a