network/shill/fixture.go: wait for connected service
This change modifies the resetShill function to wait for a connected
network instead of a visible one. Tests were running before shill had
time to create all of the default services. This can take more time so
increase the timeouts for the fixture and waiting for service
properties.

BUG=chromium:1217151
TEST=tast run fizz-dut network.Diag* (pass)
tast run fizz-dut network.DiagFailDNSResolverPresent.empty_name_servers

Change-Id: I214cec753551a4fab3cebff99f2d56fa611125ee